config: split common utility functions into util
Keep user-facing structures in common, and move internal helper functions
to a new util package.  Don't put them in an "internal" subdirectory
because we want to allow out-of-tree variant implementations.
